refactor: Set the Midnight continuous fee ceiling as an annual rate
maxContinuousFee was a per-second WAD rate mirroring the market's own encoding,
so a spell had to write 317097919 to mean one percent a year. It is now a uint16
in centi-basis points a year, grouped as percent_bp_cbp, so that ceiling reads
1_00_00 and the whole config speaks in rates.

The facet converts the ceiling rather than the fee, flooring, so the effective
ceiling never sits above the rate named. 1_00_00 converts to exactly the vendored
MAX_CONTINUOUS_FEE, which stays the single authority for the bound and keeps a
config at the top of the range admitting a market fee at the cap.

Behaviour is unchanged for equivalent settings: the market charges the rate times
the term, so comparing rates per second and rates per year is the same test. The
narrowing to uint16 also refuses to compile a stale per-second literal. Runtime
size drops 21 bytes to 14,666.

The fork boundary test moves off the ceiling to keep a one-wei boundary, which
also makes it the first test to exercise the conversion.

+++ src/facets/midnight/IMidnightFacet.sol
@@ -48,7 +48,8 @@ interface IMidnightFacet is IFacet {
      * @param  minSellTick      Lowest offer tick a sell may accept (fee-adjusted at call time).
      * @param  minBuyYield      Lowest implied yield a buy may accept (basis points a year).
      * @param  maxSellYield     Highest implied yield a sell may give up (basis points a year).
-     * @param  maxContinuousFee Highest market continuous fee a buy tolerates (per second, WAD).
+     * @param  maxContinuousFee Highest market continuous fee a buy tolerates (centi-basis points
+     *                          a year).
      * @param  maxLossFactor    Highest market loss factor a buy tolerates (fraction of
      *                          `type(uint128).max`).
      */
@@ -57,7 +58,7 @@ interface IMidnightFacet is IFacet {
         uint16  minSellTick;
         uint16  minBuyYield;
         uint16  maxSellYield;
-        uint32  maxContinuousFee;
+        uint16  maxContinuousFee;
         uint128 maxLossFactor;
     }
 
@@ -80,7 +81,8 @@ interface IMidnightFacet is IFacet {
      * @param  minSellTick      Lowest offer tick a sell may accept.
      * @param  minBuyYield      Lowest implied yield a buy may accept (basis points a year).
      * @param  maxSellYield     Highest implied yield a sell may give up (basis points a year).
-     * @param  maxContinuousFee Highest market continuous fee a buy tolerates.
+     * @param  maxContinuousFee Highest market continuous fee a buy tolerates (centi-basis points
+     *                          a year).
      * @param  maxLossFactor    Highest market loss factor a buy tolerates.
      */
     event MidnightMarketConfigSet(
@@ -89,7 +91,7 @@ interface IMidnightFacet is IFacet {
         uint16          minSellTick,
         uint16          minBuyYield,
         uint16          maxSellYield,
-        uint32          maxContinuousFee,
+        uint16          maxContinuousFee,
         uint128         maxLossFactor
     );
